backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
google-generativeai==0.8.3
langchain-core==0.3.18
langchain-community==0.3.7
faiss-cpu==1.8.0.post1
pymupdf==1.24.13
pdfplumber==0.11.4
camelot-py[base]==0.11.0
tabula-py==2.10.0
pandas==2.2.3
python-dotenv==1.0.1
markdown==3.7
pydantic-settings==2.6.1
